支払い情報をログに保存するようにした

+++ entrance-system/internal/model/payment.go
@@ -11,6 +11,6 @@ type PaymentLog struct {
 	UserID      int                `json:"user_id" bson:"user_id"`
 	Time        time.Time          `json:"time" bson:"time"`
 	Description string             `json:"description" bson:"description"`
-	Amount      float64            `json:"amount" bson:"amount"`
+	Amount      int                `json:"amount" bson:"amount"`
 	Payway      string             `json:"payway" bson:"payway"`
 }
